refactor: Move functions to util module
