+++ src/trajectory/__main__.py
@@ -1,33 +1,9 @@
 #!/usr/bin/env python
 import os
-
 import sys
 
-
-def print_usage():
-    print('Usage:')
-    print('python -m trajectory trajectory_name')
-    print('\tShows a plot of trajectory_name.')
-    print('python -m trajectory --trajectories')
-    print('\tShows the list of trajectories.')
-
-
-def print_error_message():
-    print('Error using module trajectory!')
-    print('Try python -m trajectory --help for more information.')
-
-
-def print_files_in_path(path):
-    print('Available trajectories:')
-    for trajectory in sorted(get_trajectories_in_path(path)):
-            print(trajectory)
-
-
-def get_trajectories_in_path(path):
-    return [file.replace('_trajectory.py', '')
-            for file in os.listdir(path)
-            if '_trajectory' in file and '.pyc' not in file]
-
+from .util import print_error_message, print_usage
+from .util import print_files_in_path, get_trajectories_in_path
 
 def plot_trajectory():
     pass
@@ -55,4 +31,3 @@ if __name__ == '__main__':
         sys.exit(2)
 
     plot_trajectory()
-
